+++ packages/openapi-ts/src/plugins/@hey-api/sdk/typeOptions.ts
@@ -91,6 +91,7 @@ export const createTypeOptions = ({
       ? [
           tsc.typeParameterDeclaration({
             constraint: tsc.typeReferenceNode({ typeName: 'Composable' }),
+            defaultType: tsc.typeNode("'$fetch'"),
             name: 'TComposable',
           }),
           tsc.typeParameterDeclaration({